[FormBundle] Fix namespace + modernize old fashioned mocks and use codeception
