--- src/Kunstmaan/FormBundle/Tests/Entity/FakePage.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Kunstmaan\FormBundle\Tests\Entity;
-
-use Kunstmaan\NodeBundle\Entity\AbstractPage;
-
-class FakePage extends AbstractPage
-{
-    public function getPossibleChildTypes()
-    {
-        return [];
-    }
-}
