--- src/Kunstmaan/FormBundle/Tests/Stubs/TestRepository.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Kunstmaan\FormBundle\Tests\Stubs;
-
-/**
- * TestRepository
- */
-class TestRepository
-{
-
-    /**
-     * @return NULL
-     */
-    public function find()
-    {
-        return null;
-    }
-
-}
