+++ src/Kunstmaan/FormBundle/Tests/_support/UnitTester.php
@@ -1,5 +1,5 @@
 <?php
-namespace FormBundle;
+namespace Kunstmaan\FormBundle\Tests;
 
 /**
  * Inherited Methods
