+++ src/Kunstmaan/FormBundle/Tests/_support/_generated/UnitTesterActions.php
@@ -1,5 +1,5 @@
-<?php  //[STAMP] 78ee096a68f63bd6a3edb6e7ad138714
-namespace FormBundle\_generated;
+<?php  //[STAMP] 695c6e0a30938a80cdf3e646d5614366
+namespace Kunstmaan\FormBundle\Tests\_generated;
 
 // This class was automatically generated by build task
 // You should not change it manually as it will be overwritten on next build
