+++ src/Kunstmaan/FormBundle/Tests/unit/Event/SubmissionEventTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Tests\Kunstmaan\FormBundle\Event;
+namespace Kunstmaan\FormBundle\Tests\Event;
 
-use Kunstmaan\FormBundle\Tests\Entity\FormPage;
+use Codeception\Stub;
+use Kunstmaan\FormBundle\Entity\AbstractFormPage;
 use Kunstmaan\FormBundle\Entity\FormSubmission;
 use Kunstmaan\FormBundle\Event\SubmissionEvent;
 
@@ -15,15 +16,15 @@ class SubmissionEventTest extends \PHPUnit_Framework_TestCase
     public function testEvent()
     {
         $submission = new FormSubmission();
-        $page = new FormPage();
+        $page = Stub::makeEmpty(AbstractFormPage::class, [
+            'getId' => 2
+        ]);
 
         $submission->setId(1);
-        $page->setId(2);
 
         $event = new SubmissionEvent($submission, $page);
         $this->assertInstanceOf(FormSubmission::class, $event->getSubmission());
-        $this->assertInstanceOf(FormPage::class, $event->getPage());
+        $this->assertEquals($page, $event->getPage());
         $this->assertEquals(1, $event->getSubmission()->getId());
-        $this->assertEquals(2, $event->getPage()->getId());
     }
 }
