+++ src/Kunstmaan/FormBundle/Tests/unit/Form/EmailFormSubmissionTypeTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Kunstmaan\FormBundle\Form;
+namespace Kunstmaan\FormBundle\Tests\Form;
 
 use Kunstmaan\FormBundle\Entity\FormSubmissionFieldTypes\EmailFormSubmissionField;
 use Kunstmaan\FormBundle\Form\EmailFormSubmissionType;
