+++ src/Kunstmaan/FormBundle/Tests/unit/Resources/assets/example.jpg
